+++ src/migrations/index.js
@@ -8,7 +8,6 @@ const migrations = [
   require("./007_animal_microchip_details"),
   require("./008_vaccination_presets"),
   require("./009_repair_vaccination_presets"),
-  require("./010_care_management"),
   require("./011_disposal_facilities"),
 ];